bwtonqr/requirements.txt
qrcode==6.1
Image==1.5.28
pika==1.1.0
oss2
python-dotenv
pylint
mypy